More work on unifying the types used throughout the API.  Also fixes many compiler warnings on GCC when options -Wconversion and -Werror are added to the configure stage.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@925834 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/wireformat/openwire/marshal/BaseDataStreamMarshaller.cpp
@@ -694,7 +694,7 @@ std::vector<unsigned char> BaseDataStreamMarshaller::tightUnmarshalByteArray(
         if( bs->readBoolean() ) {
             int size = dataIn->readInt();
             data.resize( size );
-            dataIn->readFully( &data[0], data.size() );
+            dataIn->readFully( &data[0], (int)data.size() );
         }
 
         return data;
@@ -715,7 +715,7 @@ std::vector<unsigned char> BaseDataStreamMarshaller::looseUnmarshalByteArray(
             int size = dataIn->readInt();
             std::vector<unsigned char> data;
             data.resize( size );
-            dataIn->readFully( &data[0], data.size() );
+            dataIn->readFully( &data[0], (int)data.size() );
             return data;
         }
 
@@ -736,7 +736,7 @@ std::vector<unsigned char> BaseDataStreamMarshaller::tightUnmarshalConstByteArra
     try{
         std::vector<unsigned char> data;
         data.resize( size );
-        dataIn->readFully( &data[0], data.size() );
+        dataIn->readFully( &data[0], (int)data.size() );
         return data;
     }
     AMQ_CATCH_RETHROW( IOException )
@@ -753,7 +753,7 @@ std::vector<unsigned char> BaseDataStreamMarshaller::looseUnmarshalConstByteArra
     try{
         std::vector<unsigned char> data;
         data.resize( size );
-        dataIn->readFully( &data[0], data.size() );
+        dataIn->readFully( &data[0], (int)data.size() );
         return data;
     }
     AMQ_CATCH_RETHROW( IOException )
