+++ src/main/decaf/util/zip/Inflater.cpp
@@ -340,7 +340,7 @@ int Inflater::inflate( unsigned char* buffer, int size, int offset, int length )
                 __FILE__, __LINE__, "length parameter out of Bounds: %d.", length );
         }
 
-        int outStart = this->data->stream->total_out;
+        ulong outStart = this->data->stream->total_out;
 
         this->data->stream->next_out = buffer + offset;
         this->data->stream->avail_out = (uInt)length;
@@ -363,7 +363,7 @@ int Inflater::inflate( unsigned char* buffer, int size, int offset, int length )
                 __FILE__, __LINE__, "Inflate failed because a block of invalid data was found." );
         }
 
-        return this->data->stream->total_out - outStart;
+        return (int)( this->data->stream->total_out - outStart );
     }
     DECAF_CATCH_RETHROW( NullPointerException )
     DECAF_CATCH_RETHROW( IndexOutOfBoundsException )
